requirements.txt
aiohttp==3.6.2
python-socketio==4.5.1
